requirements.txt
Pillow
MoviePy